objects integer conv. to long

+++ src/main/java/com/fhdw/objects/Studiengruppe.java
@@ -17,7 +17,7 @@ public class Studiengruppe {
     @Id
     @GeneratedValue(strategy=GenerationType.AUTO)
     private Long StudiengruppenID;
-    private Integer MentorID;
+    private Long MentorID;
     private String StudiengruppeName;
     private String StudiengangName;
     private String AbschlussTitel;
@@ -28,7 +28,7 @@ public class Studiengruppe {
 
     protected Studiengruppe(){};
 
-    public Studiengruppe(Long studiengruppenID, Integer mentorID, String studiengruppeName, String studiengangName, String abschlussTitel, Date startDatum, Date endeDatum) {
+    public Studiengruppe(Long studiengruppenID, Long mentorID, String studiengruppeName, String studiengangName, String abschlussTitel, Date startDatum, Date endeDatum) {
         StudiengruppenID = studiengruppenID;
         MentorID = mentorID;
         StudiengruppeName = studiengruppeName;
@@ -46,11 +46,11 @@ public class Studiengruppe {
         StudiengruppenID = studiengruppenID;
     }
 
-    public Integer getMentorID() {
+    public Long getMentorID() {
         return MentorID;
     }
 
-    public void setMentorID(Integer mentorID) {
+    public void setMentorID(Long mentorID) {
         MentorID = mentorID;
     }
